refactor: AwsS3Service -> ImageService 로 변경
