+++ src/main/java/com/donation/domain/post/application/Image/ImageService.java
@@ -1,4 +1,4 @@
-package com.donation.infrastructure.Image;
+package com.donation.domain.post.application.Image;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class AwsS3Service {
+public class ImageService {
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
     private final AmazonS3 amazonS3;
