+++ src/main/java/com/donation/presentation/PostController.java
@@ -1,7 +1,7 @@
 package com.donation.presentation;
 
 import com.donation.domain.post.dto.*;
-import com.donation.domain.post.service.PostService;
+import com.donation.domain.post.application.PostService;
 import com.donation.infrastructure.common.CommonResponse;
 import com.donation.infrastructure.util.CursorRequest;
 import com.donation.infrastructure.util.PageCursor;
