+++ src/test/java/com/donation/domain/user/application/UserServiceTest.java
@@ -1,18 +1,17 @@
 package com.donation.domain.user.application;
 
-import com.donation.domain.user.service.UserService;
-import com.donation.presentation.auth.LoginMember;
-import com.donation.domain.user.dto.UserLoginReqDto;
-import com.donation.domain.user.dto.UserSaveReqDto;
-import com.donation.domain.user.dto.UserEmailRespDto;
-import com.donation.domain.user.dto.UserRespDto;
 import com.donation.common.utils.ServiceTest;
-import com.donation.domain.user.entity.User;
-import com.donation.global.exception.DonationInvalidateException;
-import com.donation.domain.user.repository.UserRepository;
-import com.donation.infrastructure.util.PageCustom;
 import com.donation.domain.auth.application.AuthService;
-import com.donation.infrastructure.Image.AwsS3Service;
+import com.donation.domain.user.dto.UserEmailRespDto;
+import com.donation.domain.user.dto.UserLoginReqDto;
+import com.donation.domain.user.dto.UserRespDto;
+import com.donation.domain.user.dto.UserSaveReqDto;
+import com.donation.domain.user.entity.User;
+import com.donation.domain.user.repository.UserRepository;
+import com.donation.global.exception.DonationInvalidateException;
+import com.donation.domain.post.application.Image.ImageService;
+import com.donation.infrastructure.util.PageCustom;
+import com.donation.presentation.auth.LoginMember;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -37,7 +36,7 @@ class UserServiceTest extends ServiceTest {
     private UserRepository userRepository;
 
     @Autowired
-    private AwsS3Service awsS3Service;
+    private ImageService imageService;
 
     @Autowired
     private AuthService authService;
@@ -75,7 +74,7 @@ class UserServiceTest extends ServiceTest {
         assertThat(actual.getProfileImage()).isNotEqualTo(일반_사용자_프로필);
 
         //clear
-        awsS3Service.delete(actual.getProfileImage());
+        imageService.delete(actual.getProfileImage());
     }
 
     @Test
