+++ src/test/java/com/donation/presentation/PostControllerTest.java
@@ -4,7 +4,7 @@ import com.donation.common.utils.ControllerTest;
 import com.donation.domain.auth.application.AuthService;
 import com.donation.domain.post.dto.PostListRespDto;
 import com.donation.domain.post.entity.PostState;
-import com.donation.domain.post.service.PostService;
+import com.donation.domain.post.application.PostService;
 import com.donation.global.exception.DonationNotFoundException;
 import com.donation.infrastructure.util.CursorRequest;
 import com.donation.infrastructure.util.PageCursor;
